Fix fetcher test.
