+++ apps/explorer/config/config.exs
@@ -36,7 +36,6 @@ config :explorer, Explorer.Celo.AbiHandler, enabled: true
 
 config :explorer, Explorer.Celo.SignerCache, enabled: true
 
-config :explorer, Explorer.Celo.CoreContracts, enabled: true, refresh: :timer.hours(1)
 
 config :explorer, :stacktrace_depth, 20
