+++ apps/explorer/config/dev.exs
@@ -39,6 +39,8 @@ variant =
     |> String.downcase()
   end
 
+config :explorer, Explorer.Celo.CoreContracts, enabled: true, refresh: :timer.hours(1)
+
 # Import variant specific config. This must remain at the bottom
 # of this file so it overrides the configuration defined above.
 import_config "dev/#{variant}.exs"
