+++ apps/explorer/config/test.exs
@@ -60,6 +60,9 @@ variant =
     |> String.downcase()
   end
 
+
+config :explorer, Explorer.Celo.AddressCache, Explorer.Celo.AddressCache.Mock
+
 # Import variant specific config. This must remain at the bottom
 # of this file so it overrides the configuration defined above.
 import_config "test/#{variant}.exs"
