+++ apps/indexer/test/indexer/block/realtime/fetcher_test.exs
@@ -63,7 +63,8 @@ defmodule Indexer.Block.Realtime.FetcherTest do
     } do
       celo_token_address = insert(:contract_address)
       insert(:token, contract_address: celo_token_address)
-      "0x" <> unprefixed_celo_token_address_hash = to_string(celo_token_address.hash)
+
+      Mox.stub(Explorer.Celo.AddressCache.Mock, :contract_address, fn _name -> to_string(celo_token_address.hash) end)
 
       {:ok, sequence} = Sequence.start_link(ranges: [], step: 2)
       Sequence.cap(sequence)
@@ -236,17 +237,6 @@ defmodule Indexer.Block.Realtime.FetcherTest do
                  "transactionHash" => "0xd3937e70fab3fb2bfe8feefac36815408bf07de3b9e09fe81114b9a6b17f55c8",
                  "transactionIndex" => "0x0"
                }
-             }
-           ]}
-        end)
-        # read_addresses for 4 smart contracts in the fetcher
-        |> expect(:json_rpc, 4, fn [%{id: id, method: "eth_call"}], _ ->
-          {:ok,
-           [
-             %{
-               jsonrpc: "2.0",
-               id: id,
-               result: "0x000000000000000000000000" <> unprefixed_celo_token_address_hash
              }
            ]}
         end)
